requirements.txt
opencv-python
mediapipe
pyautogui
screen-brightness-control
pycaw
comtypes
pillow
ttkbootstrap